+++ CHANGELOG.md
@@ -116,6 +116,23 @@
 - The language server no longer shows completions when inside a literal string.
   ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
 
+- LSP can now suggest a code action to convert `let assert Ok` into a case expression:
+
+  ```
+  let assert Ok(foo) = bar()
+  ```
+
+  Becomes:
+
+  ```
+  let foo = case bar() {
+    Ok(foo) -> foo
+    Error(e) -> panic as "value was not ok"
+  }
+  ```
+
+  ([Gears](https://github.com/gearsdatapacks))
+
 ### Bug Fixes
 
 - Fixed a bug where the compiler would report errors for duplicate `@external`